+++ src/transport/AqueousTransport.cpp
@@ -325,8 +325,8 @@ void AqueousTransport::getThermalDiffCoeffs(doublereal* const dt)
  *                        length = ldx * ndim
  */
 void AqueousTransport::getSpeciesFluxes(size_t ndim, const doublereal* const grad_T,
-                                        int ldx, const doublereal* const grad_X,
-                                        int ldf, doublereal* const fluxes)
+                                        size_t ldx, const doublereal* const grad_X,
+                                        size_t ldf, doublereal* const fluxes)
 {
     set_Grad_T(grad_T);
     set_Grad_X(grad_X);
