+++ src/transport/MultiTransport.cpp
@@ -302,9 +302,9 @@ void MultiTransport::solveLMatrixEquation()
  *                  Flat vector with the m_nsp in the inner loop.
  *                   length = ldx * ndim
  */
-void MultiTransport::getSpeciesFluxes(size_t ndim, const doublereal* const grad_T, int ldx,
-                                      const doublereal* const grad_X,
-                                      int ldf, doublereal* const fluxes)
+void MultiTransport::getSpeciesFluxes(size_t ndim, const doublereal* const grad_T,
+                                      size_t ldx, const doublereal* const grad_X,
+                                      size_t ldf, doublereal* const fluxes)
 {
     // update the binary diffusion coefficients if necessary
     updateDiff_T();
